+++ core/models/part.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from sqlalchemy import String, Integer, ForeignKey, Boolean, func
+from sqlalchemy import String, Integer, ForeignKey, Boolean, func, Float
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from .base import Base
@@ -50,8 +50,8 @@ class Part(Base, Life, Stage, CanDone):
         unique=False
     )
 
-    quantity: Mapped[int] = mapped_column(
-        Integer,
+    quantity: Mapped[float] = mapped_column(
+        Float,
         nullable=False,
         unique=False
     )
